refactor: fix IO performance issue
By default Google GMS Task library run task listener on Main Thread, this cause a lot of issue when we launch many request.

This commit implement a Thread Pool which specialize in IO intensive task.

Currently the application still laggy when open a story with close to thousand of comment. We still have many things to do.