fix: fixing weird hacker news API bug
Sometime the API will return null for specific item for no reason. If we visit the website, the comment is actually there.
Also it feels like the null item will change from time to time.

When we retrieve such null item, we doing some extra work to recover some id info, and replace it with a custom empty comment. At least from now on it won't crash our app.